Added skeleton for BSD FFS. Added HFS+. If HFS detects it has wrapped a HFS+ volume, inform that it's not its filesystem. Added a workaround in calculation with UInt32 and ulong that made HFS and MFS show a bad date.
git-svn-id: svn://claunia.com/FileSystemIDandChk@2 17725271-3d32-4980-a8cb-9ff532f270ba

+++ FileSystemIDandChk/Plugins/AppleMFS.cs
@@ -218,8 +218,8 @@ namespace FileSystemIDandChk.Plugins
 		private struct MFS_MasterDirectoryBlock // Should be offset 0x0400 bytes in volume
 		{
 			public UInt16 drSigWord;  // Signature, 0xD2D7
-			public UInt32 drCrDate;   // Volume creation date
-			public UInt32 drLsBkUp;   // Volume last backup date
+			public ulong drCrDate;   // Volume creation date
+			public ulong drLsBkUp;   // Volume last backup date
 			public UInt16 drAtrb;     // Volume attributes
 			public UInt16 drNmFls;    // Volume number of files
 			public UInt16 drDirSt;    // First directory block
